refactor: 도전자 이름 목록 객체 생성하여 Players의 역할 분리
