+++ src/test/java/blackjack/domain/player/PlayersTest.java
@@ -15,27 +15,6 @@ import org.junit.jupiter.api.Test;
 class PlayersTest {
 
     @Test
-    @DisplayName("플레이어의 이름이 중복되면 예외가 발생한다")
-    void checking_player_name_duplicated() {
-        assertThrows(DuplicatedPlayerNameException.class,
-                () -> Players.from(List.of("pobi", "pobi")));
-    }
-
-    @Test
-    @DisplayName("도전자의 숫자가 10명이 초과되면 예외가 발생한다.")
-    void over_challenger_number() {
-        assertThrows(InvalidChallengerNumberException.class,
-                () -> Players.from(List.of("1", "2", "3", "4", "5", "6", "7", "8", "9", "10", "11")));
-    }
-
-    @Test
-    @DisplayName("도전자의 숫자가 1명 미만이면 예외가 발생한다.")
-    void under_challenger_number() {
-        assertThrows(InvalidChallengerNumberException.class,
-                () -> Players.from(Collections.emptyList()));
-    }
-
-    @Test
     @DisplayName("challenger만 반환하는지 테스트")
     void return_challengers() {
         Players players = Players.from(List.of("pobi", "oing"));
